+++ src/MiniTone.h
@@ -14,7 +14,7 @@ class MiniTone {
 public:
   void begin(uint8_t buzzer_pin);
   void play(uint8_t count, ...);
-  void play(uint16_t frequency, uint32_t duration);
+  void play(uint16_t frequency, uint16_t duration);
 private:
   uint8_t _buzzer_pin;
 };
